Add headless Service for SwiftStorage
Expose the storage backend services as a headless network service. A
headless service was choosen for two reasons: load balancing is not
required for the backend services and k8s will setup DNS entries that
point directly to the individual pods. These DNS entries will then be
used in the Swift rings to setup and connect devices.