Create ConfigMap using template conf files for SwiftStorage
Provide minimum default config for the SwiftStorage controller using a
static, default swift.conf. Hash prefixes/suffixes in the latter one
should be provided using a Secret for production environments in the
future.

Backend service configuration includes a workaround to prevent errors
when connecting to (the non-existing) /dev/log. This currently raises an
exception, thus using UDP logging on startup, and then switching to
verbose mode and logging to the console afterwards. This has been fixed
upstream in later releases.