Add Swift Ring controller
This is a very early implementation of the SwiftRing controller. For now
it only creates initial rings to be used by the storage and proxy
services. Rings are pretty static and created only based on the number
of devices given.

Besides the initial device addition and rebalancing no modificiation is
done throughout the lifecycle at this point. More work is required to
update rings (add & remove devices) and rebalance as needed.

The required Python scripts are run using a Job. Rings are sre stored in
a ConfigMap that is created within the Job itself using the service
account.