Use Secrets to store sensitive data
swift_hash_path_* values in swift.conf should be kept private, thus
storing them in a Secret. This Secret must exist before, otherwise it
will be created with random values.

The service password for Keystone should be kept private as well, thus
using a Secret instead of a ConfigMap to provide config data for the
Proxy.