Use devices from StatefulSet to create Swift rings
Once all SwiftStorage instances are up, a list of devices with their
actual size will be created and stored in a ConfigMap. This ConfigMap is
then used by the SwiftRing instance to update the rings.

The settings StoragePodPrefix, StorageServiceName and Devices are no
longer used, removing them from CRD and samples therefor.