Create Dummy rings
Creating Swift rings is a chicken and egg problem: to create the rings
infos about the devices like hostname and sizes of the attached PVCs are
required, but to get the data we need the storage StatefulSet to be
running - which itself requires valid Swift rings, otherwise Swift
processes will crash.

This is now done by creating initial dummy rings, which basically
contain a single host. These ring files are not really usable, but
sufficient to start the container processes.

The ConfigMap for the Swift rings will be created empty initially to
allow mounting before rings are created. It will be updated later on,
and the ring-sync container will copy the real ring files to their
correct place.