Refactor devicelist function
The function creates a CSV list of devices based on the actual PVCs used
by the SwiftStorage StatefulSet. However, this is only known after
running with all required replicas, which itself is late due to the
requirement for Swift rings in advance.

This refactoring will also generate a CSV list even if the PVCs do not
exist yet; in that case it will just use the requested capacity as
value. It will be called again after the StatefulSet is fully up and
update the weights with the actual PVC site.