Refactor ring rebalance script
The script will now fetch the ring ConfigMap directly and either create
or patch it at the end. This removes the requirement to mount the
ConfigMap itself beforehand, which itself no longer requires an (empty)
ConfigMap to exist.

Note that this is just an initial POC-kind of script; this will be
refactored and more logic will be added.