Do not check if ring ConfigMap exists
Instead of checking if the ring ConfigMap does exist, just continue with
the deployment. ConfigMaps are attached as volumes to the pods and these
will wait until they do exist.

This is also possible because there are no more "empty" ConfigMaps, thus
a check for specific keys is also no longer required.