Remove creating of dummy rings in swift-init.sh
ConfigMaps are limited to 1 MiB in size, and very large deployments with
hundreds of nodes might require a few MiBs for all ring and builder
files. These files will be split into multiple ConfigMaps in a follow
up, and need to be concatenated before starting service processes. This
logic will be added in a follow up to the init script.