Remove InitContainers
The ring-sync container actually does everything that is required, there
is no need for InitContainers anymore.

The order of the containers within the pods has been changed, the
ring-sync containers are now first to give them a chance to execute once
before other services are started. There is no guarantee about this, but
failures are tolerated and recovered from.