[tlse] tls for SwiftProxy pod configuration
Public/Internal service cert secrets and the CA bundle secret
can be passed to configure httpd virtual hosts for tls termination.
The certs get direct mounted to the appropriate place in
etc/pki/tls/certs/%s.crt|key and a CA bundle to
/etc/pki/ca-trust/extracted/pem/tls-ca-bundle.pem . Job deployments
for bootstrap/cron get the CA bundle added if configured.

Also indexes the named input resources for password, CA bundle,
and endpoint secrets to be able to watch them for any changes and
reconcile if needed.

Note: Atm the CA bundle is not being attached to SwiftStorage or
SwiftRing deployments.

Depends-On: https://github.com/openstack-k8s-operators/lib-common/pull/428

Jira: OSPRH-4371