Add DNSData entries for SwiftStorage service pods
Storage service hostnames need to be resolvable on the dataplane nodes
as well. They are used within the Swift rings to be able to allow node
IP changes without updating and redistributing Swift rings.

This patch creates DNSData entries for every storage service pod,
similar to [1]. It only creates DNS for the storage network, as this is
the only one that should be used within the storage backend services.

There is no delete function to scale down, as we don't support scaling
down for SwiftStorage instances. However, the created DNSData CRs are
owned by the SwiftStorage instance, thus being deleted properly if the
instance is deleted.

[1] https://github.com/openstack-k8s-operators/ovn-operator/pull/154