Mount dispersion.conf in rebalance job
Optional mount because dispersion.conf does not exist when the
deployment is just starting. The file is created by the SwiftProxy
controller later on.

The file is not required for the deployment itself, but only for
administrators that want to run the dispersion-tools later on manually.

At the same time directly mounting swift.conf directly, thus no copying
is required anymore.