Avoid second rebalance on startup
Until now the SwiftRing instance created a ring with estimated disk
sizes if PVCs and size data was not yet available. Rings were updated
later once sizes were known.

However, this is no longer needed. PVCs have to be created eventually if
SwiftStorage replicas > 0, thus we can requeue until these are bound.
This eliminates a second rebalance with the correct values shortly
after.

It also simplifies the rebalance script and removes a possible risk when
changing weights for existing devices. Initially set weights are kept
now, which should be the correct ones from the beginning.