Improve logging output of ring-sync.sh
The script is used by all Swift containers, but not every config file is
needed or available in all cases. Adding checks to avoid warnings in the
logs that might be misleading, eg. if directories exist or files do not.

Also be less verbose during normal executions, but get a list of all
files in /etc/swift to help debugging if needed.