Verify network attachments after all pods are ready
Verify network attachments before all pods are ready will always report
an error and wrong status.

In case of the proxy it should also be checked that the number of active
replicas is equal to the requested size - the deployment is already
usable with a single proxy replica, but not matching the requested size.