Fix regression when scaling up
The default PodManagementPolicy for StatefulSet is set to
OrderedReadyPodManagement, which means it only creates new pods when the
previous one is ready.

However, to start the pods the Swift ring file ConfigMap is needed, and
this in turn is only created once all requested pods are started and
their PVCs are bound (because it needs information on the PVCs to be
used). If the instance is scaled by more than 1 (which is the case also
for starting with 3 replicas), it failed.

This is a regression to previous behaviour, introduced by changes to the
ring setup. Previously the ring rebalance did not wait until all PVCs
where bound, but estimated their size and updated the rings afterwards
with the correct size. This required ring rebalances and data movements
even for new deployments.

Therefore changing the policy to ParallelPodManagement, wich starts the
number of requested pods immediately. This does only affect scaling, not
updates.

The kuttl tests has been updated to avoid regressions in the future.