Add replication kuttl test
This is basically a full deployment test, ensuring all parts work
together as expected:

- Create a 3x replicated deployment
- Upload some data and ensure it is accessible
- Forcefully delete one replica on the disk
- Ensure replication fixes this and gets back to 3 replicas

It uses the swift-dispersion tools and user, which creates objects
distributed across the nodes and fetches them later again to ensure
they are replicated properly.

This kuttl uses all important parts of the Swift deployment:
- Swift proxy using Keystone for authentication
- network connectivity between proxy and storage services
- replication between storage services

The test takes approx 2.5 minutes of time to finish.