Use projected volumes for ring and config files
Manually merging files and checking ring updates using a helper script
is no longer required with this change. This is a much cleaner approach,
reduces the complexity and uses common k8s patterns.

Updating the ring ConfigMap and its files is detected by the Swift
services, eventually re-reading the ring files.

This change impacts the use of the defaultConfigOverwrite option.
ConfigMap keys must be known in advance, thus adding a custom key for
every config with a dummy template file.