Do not create new builder file automatically if ring.gz does exist
This might be on purpose if an operator created the CM and manages these
manually. In that case it would be harmful to create a new .builder file
and overwriting the .ring.gz file eventually.

Fixing an if-condition along the way; the swiftrings.tar.gz file would
be empty if the CM key does not exist due to redirecting the output and
raising an error.