Change PVC prefix, CM name and PV mountpoint name
The "srv" prefix for the PVCs is replaced with "swift" to align with
other PVC names and making it clear which service it is used by, even
if the CR does use a non-default name.

The storage-devices CM name does not need to be hardcoded, changing this
and just using the instance name with a "-config-data" suffix.

Last but not least the PV mountpoint has been renamed from d1. This was
the old name used by TripleO, now making it more clear that this is a PV
mountpoint. This is especially helpful during dataplane adoption; old
devices on the controller nodes use "d1", new devices "pv". This helps
to read the output of swift-ring-builder commands.