Do not use MustParse to process PVC size
The resource.MustParse function was previously used by the operator to
process the storageSize request provided by the top-level Swift CR.
However, if the human operator passes a wrong StorageRequest size, a
panic() is generated without any chance to recover. This happens because
the MustParse implementation wraps the ParseQuantity function, but
it returns panic() instead of propagating the error that can be easily
caught at the operator level. This change moves away from the MustParse
usage and rely on the ParseQuantity mechanism, providing an error that
can be handled by the operator.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>