Remove explicit instance Update calls
Some of those calls for newInstance was directly redundant as the
self finalizer condition already handled the newInstance case as well.
Other places used explicit status updates instead of relying on the
deferred PatchInstance call. In general explicit update is dangerous
as it only updates either the main resource or the status subresource,
but then the result of the API call is used to update the instance
variable overriding any not persisted in memory updates.