Allow to scale down SwiftStorage Pods
If swift is not explicitly created in the ctlplane (enabled: false), we
still get it as part of the OpenStackControPlane, and swiftStorage has
replicas: 1.
This is problematic in adoption, where if we get replica: 1 by default,
we fail later in the process when we try to only enable SwiftProxy and
pass replicas: 0 to SwiftStorage (the Webhook denies scale down).
Given that SwiftStorage is a StatefulSet, we preserve network and
storage properties, and it might be possible that we want to simply
turn off an instance for troubleshooting purposes, so this patch removes
the webhook check.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>