Fix race condition if ring rebalance is needed
The hashes to trigger a new run of the rebalance job could actually be
set even without deleting the job (for example if the existing one was
still active). This commit changes the behaviour, only resetting the
hashes if the job got either successfully deleted or did not exist at
all.