Fix race condition if ring rebalance is needed
If a rebalance is needed and the job has to re-run, it will be deleted
first. Deletes are executed using a background deletion policy, thus a
deleted job might still exist and if the operator tries to re-create it,
it will immediately become succesful (and not executed at all).

This commit changes the behaviour to reconcile again if the job does
exist.