Move to VerifySecret when checking the ctlplane secret
Currently we get the osp_secret and we build the hash of the entire
secret. If an unrelated field is added or changed, this cause a rollout
of new swiftProxy Pods. This change moves the GetSecret function call,
used to compute the hash and setting it in the envVars, to VerifySecret:
it expects to check only the expectedFields passed as input and it's
consistent with the work already done in Glance, Cinder, and Manila.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>