Add Topology kuttl tests
This patch adds a new step within the existing "customization" kuttl
test to reference a new topology and assert that both swift-proxy and
swift-storage properly apply the referenced topology.
By doing this we can ensure that the reference is processed, the
condition is marked as true, and the status.lastAppliedTopology is
populated with the name of the topology.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>