Add node-awareness to ring rebalance when using PVs
The ring rebalance uses the hostname/IP of the device to distribute
replicas across nodes. However, since there is one unique hostname per
pod and multiple pods might be running on the same hardware node, this
does not guarantee replica assignment to different nodes when using PVs
as backend and using multiple PVs per node.

To ensure devices are distributed within the cluster as far from each
other as possible, the underlying nodename needs to be used when
rebalancing.

The node name of each PVs is already stored within the metadata field of
the Swift ring. This change makes use of this when rebalancing: it swaps
the IP and meta fields, rebalances (now using the actual node name
itself and not the pod name) and afterwards swaps back meta and IP
field. Afterwards write_ring is called to rewrite the .ring.gz files
from the .builder files without rebalancing again.

Jira: OSPRH-6928